+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 # PayPal iOS SDK Release Notes
 
 ## unreleased
-* `Card`:
+* Card
   * Remove `ThreeDSecureRequest` from `CardRequest` and create URLs internally 
   * Update `CardRequest` to optionally pass `sca`
-* PayPalUI:
+  * Remove step requiring `ASWebAuthenticationPresentationContextProviding` conformance
+* PayPalWebCheckout
+  * Remove step requiring `ASWebAuthenticationPresentationContextProviding` conformance
+* PayPalUI
   * Fix assets not rendering when importing with Swift Package Manager
 
 ## 0.0.3 (2022-11-03)